Big Update
Currency Auto-pull is in and working. got the First two opportunities down for the three routes that we have so far. Going to focus on Mole & joining family routes first. Added scripts for the rest of the verbs and fixed there snap in areas. Added zoom to the camera. Added TextMeshPro to the project but have not used or tried it properly yet. Added card-prefabs for the cards I needed. Added a list to easily instntiate cards when needed. Need to make a very simple menu system for choosing a starting class and instantiating their cards before the game starts. Have some of the main. ellements in but not cohesive or together yet.