Multi-Slot and Multi-Card Support Update
Reworked the verb slots and made them able to handle multiple cards based on on need. Also added a currency counter text to ingame UI. Updated card tracking while in the slots. Also updated VerbTimer so that card spawns and destructions happen after the collect button is pressed. Still some kinks to work out and refine but better than it was and I am able to rework and move forward with what we had planned now.