Big Update
Added a lot of information. Added temporary endings to the undeveloped routes. Fixed the feedback on the undeveloped routes. Added text info to the card info boxes in the slot areas. Speaking with skills spawns people cards that are used during some opportunities. Can currently have multiple of the same card, but will chnage in the future as I am too tired to do so currently. This is right before making a build.